requirements.txt
livekit-agents[deepgram,google,cartesia,silero,turn-detector]~=1.2
python-dotenv>=1.0.1
aiohttp>=3.9.0
requests>=2.31.0